fix rebase issues
