+++ torch_xla/csrc/runtime/initialize_pjrt.h
@@ -1,0 +1,15 @@
+#ifndef XLA_CLIENT_INITIALIZE_PJRT_CLIENT_H_
+#define XLA_CLIENT_INITIALIZE_PJRT_CLIENT_H_
+
+
+#include "xla/pjrt/pjrt_client.h"
+
+namespace torch_xla {
+namespace runtime {
+
+std::unique_ptr<xla::PjRtClient> InitializePjRt(const std::string& device_type);
+
+}
+}
+
+#endif // XLA_CLIENT_INITIALIZE_PJRT_H_
